feat: cars filter

+++ controllers/cars.route.js
@@ -56,11 +56,41 @@ router.get('/view/:carId', showCarById);
 
 
 async function listAllCars(req, res){
-    const cars = await carRepository.getAllCars();
-    res.render("cars/list", {cars: cars});
+    // check query: transmission=&passengers=&energyType=1 if types are ok
+    console.log(req.query);
+    if(req.query.transmission && (req.query.transmission != 'true' && req.query.transmission != 'false')){
+        req.session.errorMessage = "Invalid transmission type";
+        res.redirect("/cars/list");
+        return;
+    }
+    if(req.query.passengers && (req.query.passengers != '2' && req.query.passengers != '3' && req.query.passengers != '4' && req.query.passengers != '5' && req.query.passengers != '6' && req.query.passengers != '7')){
+        req.session.errorMessage = "Invalid passengers type";
+        res.redirect("/cars/list");
+        return;
+    }
+    if(req.query.energyType && (req.query.energyType != '1' && req.query.energyType != '2' && req.query.energyType != '3' && req.query.energyType != '4')){
+        req.session.errorMessage = "Invalid energy type";
+        res.redirect("/cars/list");
+        return;
+    }
+    let params = {
+        passengers: parseInt(req.query.passengers),
+        energyType: parseInt(req.query.energyType)
+    }
+    if(req.query.transmission){
+        params.isManual = req.query.transmission == 'true' ? 1 : 0;
+        console.log(params.isManual);
+    }
+    const cars = await carRepository.filterCars(params);
+    res.render("cars/list", {cars: cars, params: params});
 }
 
 async function showCarById(req, res){
+    if(!req.params.carId){
+        req.session.errorMessage = "No car id";
+        res.redirect("/cars/list");
+        return;
+    }
     const id = req.params.carId;
     const car = await carRepository.getCarById(id);
     const rents = await rentsRepository.getRentsByCarIdWithProfileInfos(id);
@@ -75,7 +105,13 @@ async function showCarManagement(req, res){
 
 async function createCar(req, res){
     console.log(req.body);
-    // TODO : check if all fields are filled
+
+    if(!req.body.brand || !req.body.model || !req.body.carDescription || !req.body.isManual || !req.body.passengers || !req.body.energyType || !req.body.doorsAmount || !req.body.pricePerDay){
+        req.session.errorMessage = "Please fill all fields";
+        res.redirect('/cars/manage');
+        return;
+    }
+
     if(req.file){
         console.log(req.file);
         req.body.image = `/static/img/${req.file.filename}`;
@@ -90,7 +126,13 @@ async function createCar(req, res){
 async function editCarById(req, res){
     const id = req.params.carId;
     console.log(req.body);
-    // TODO : check if all fields are filled
+
+    if(!req.body.brand || !req.body.model || !req.body.carDescription || !req.body.isManual || !req.body.passengers || !req.body.energyType || !req.body.doorsAmount || !req.body.pricePerDay){
+        req.session.errorMessage = "Please fill all fields";
+        res.redirect('/cars/manage');
+        return;
+    }
+
     if(req.file){
         console.log(req.file);
         req.body.image = `/static/img/${req.file.filename}`;
@@ -101,6 +143,11 @@ async function editCarById(req, res){
 }
 
 async function deleteCarById(req, res){
+    if(!req.params.carId){
+        req.session.errorMessage = "No car id";
+        res.redirect("/cars/manage");
+        return;
+    }
     const id = req.params.carId;
     carRepository.removeCar(id);
     res.redirect("/cars/manage");
